pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "policy-eval"
version = "0.1.0"
description = "OPA-backed workspace + patch evaluation (LLM grounding layer)"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0",
  "unidiff>=0.7.5",
  "pathspec>=0.12.0",
]

[project.scripts]
policy-eval = "policy_eval.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
